+++ src/app/group/[groupId]/solved-detail/[id]/action.ts
