--- src/app/group/[groupId]/solved-detail/[id]/quries.ts
@@ -1,9 +1,0 @@
-import { getSolution } from "@/api/solution";
-import { useSuspenseQuery } from "@tanstack/react-query";
-
-export const useSolutionQuery = (solutionId: number) => {
-  return useSuspenseQuery({
-    queryKey: ["solution", solutionId],
-    queryFn: () => getSolution(solutionId),
-  });
-};
